+++ Classes/OldModel/LegacyEvent.php
@@ -3647,4 +3647,11 @@ class LegacyEvent extends AbstractTimeSpan
             ->execute()
             ->fetchAll();
     }
+
+    public function hasTime(): bool
+    {
+        // Events with time slots should only display their time with the time slots,
+        // but not as general times in order to avoid confusion.
+        return !$this->hasTimeslots() && parent::hasTime();
+    }
 }