+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 
+- Only display the event times if there are no time slots (#3839)
 - Fix translation for flash message in BE module (#3811)
 - Ignore the webinar URL when duplicating an event (#3806)
 - Fix incorrect translation references in the TCEforms (#3801)